Fix work with user

+++ rest-api/controllers/UserController.js
@@ -18,23 +18,34 @@ class UserController {
             user = await UserService.findById(id)
         } else if (email) {
             user = await UserService.findByEmail(email)
-        }else{
+        } else {
             res.json({ message: "Enter id or email" })
         }
-        if (!user) return res.sendStatus(400)
+        if (!user) return res.status(200).json({message:"User not exists"})
         res.json(user);
     }
 
     async update(req, res) {
-        res.send("Ok")
-    }
-
-    async delete(req, res, user) {
-        const { id } = req.params;
-        if (!id) {
+        const newData = req.body.newData;
+        if (!newData._id || !req.isAuthenticated) {
             return res.sendStatus(400)
         }
-        
+
+        const updatedUser = await UserService.update(newData)
+        if (updatedUser) {
+            res.json({ updatedUser })
+        } else {
+            res.sendStatus(500)
+        }
+    }
+
+    async delete(req, res) {
+        const id = req.params.id;
+        const user = req.user;
+        if (!id || !req.isAuthenticated || user._id.toString() != id) {
+            return res.sendStatus(400)
+        }
+
         const deletedUser = await UserService.deleteById(id)
         if (deletedUser) {
             res.json({ deletedUser })
